Refactor leader and follower steps to either/or clauses
This refactoring is a major shift from how the steps had been written.
It allows for the leader and follower steps to be run in parallel for
each process. This is necessary since pluscal does not include a concept
of subprocesses. Another benefit of this change is that it brings the
pluscal spec closer to the description of the Zab algorithm that talks about
enabled steps.